gpl: converting NesterovBase::binSizeX and NesterovBase::binSizeY to double avoid rounding errors

+++ src/gpl/src/fft.cpp
@@ -42,7 +42,7 @@
 
 namespace gpl {
 
-FFT::FFT(int binCntX, int binCntY, int binSizeX, int binSizeY)
+FFT::FFT(int binCntX, int binCntY, float binSizeX, float binSizeY)
     : binCntX_(binCntX),
       binCntY_(binCntY),
       binSizeX_(binSizeX),
